This is the week 3 assignment for R Programming by JHU. It contains the solution for the makeCacheMatrix and cacheSolve functions.
This task required us to create a function that will cache the inverse of a matrix. There are already 2 functions, makeCacheMatrix and cacheSolve. We need to put a solution that will cache the inverse of a matrix. I just followed the sample format posted in the course and modified it a little in order to comply what is asked.